23w42a

+++ minecraft/src/net/minecraft/world/level/block/PointedDripstoneBlock.java
@@ -130,7 +130,11 @@ public class PointedDripstoneBlock extends Block implements Fallable, SimpleWate
 	@Override
 	public void onProjectileHit(Level level, BlockState blockState, BlockHitResult blockHitResult, Projectile projectile) {
 		BlockPos blockPos = blockHitResult.getBlockPos();
-		if (!level.isClientSide && projectile.mayInteract(level, blockPos) && projectile instanceof ThrownTrident && projectile.getDeltaMovement().length() > 0.6) {
+		if (!level.isClientSide
+			&& projectile.mayInteract(level, blockPos)
+			&& projectile.mayBreak(level)
+			&& projectile instanceof ThrownTrident
+			&& projectile.getDeltaMovement().length() > 0.6) {
 			level.destroyBlock(blockPos, true);
 		}
 	}
